LayoutTests/fast/dom/SelectorAPI/caseID-strict-expected.txt

PASS document.querySelector('#lower1').textContent is 'lower 1'
PASS document.querySelector('#LOWER2') is null
PASS document.querySelector('#UPPER1').textContent is 'UPPER 1'
PASS document.querySelector('#upper2') is null
PASS successfullyParsed is true

TEST COMPLETE
